chore: Move controller, service, dto, domain, repository each domain
