+++ src/main/java/com/ssafy/todolist/email/repository/EmailRepository.java
@@ -1,0 +1,12 @@
+package com.ssafy.todolist.email.repository;
+
+import com.ssafy.todolist.email.domain.Email;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface EmailRepository extends JpaRepository<Email, Integer> {
+
+    Email findByEmail(String email);
+
+}
