+++ src/main/java/com/ssafy/todolist/todo/domain/Todo.java
@@ -1,4 +1,4 @@
-package com.ssafy.todolist.domain.todo;
+package com.ssafy.todolist.todo.domain;
 
 import jakarta.persistence.*;
 import org.hibernate.annotations.ColumnDefault;
