+++ src/main/java/com/ssafy/todolist/todo/dto/TodoDTO.java
@@ -1,4 +1,6 @@
-package com.ssafy.todolist.domain.todo;
+package com.ssafy.todolist.todo.dto;
+
+import com.ssafy.todolist.todo.domain.Todo;
 
 public record TodoDTO(int id, String content, boolean completed) {
